DroneDelivery-RL/requirements.txt
# Core ML and RL libraries
torch>=1.13.0
numpy>=1.21.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
matplotlib>=3.5.0
scipy>=1.9.0

# AirSim integration
airsim>=1.6.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Configuration and logging
PyYAML>=6.0
tensorboard>=2.8.0
wandb>=0.13.0

# Data processing
pandas>=1.4.0
seaborn>=0.11.0

# Path planning
networkx>=2.8.0
scikit-learn>=1.1.0

# ROS2 integration (optional)
rclpy>=3.3.0
geometry_msgs>=4.2.0
sensor_msgs>=4.2.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
